Fix/update conversions and tests

+++ applications/tari_app_grpc/src/conversions/output_features.rs
@@ -54,7 +54,7 @@ impl TryFrom<grpc::OutputFeatures> for OutputFeatures {
         } else {
             Some(PublicKey::from_bytes(features.parent_public_key.as_bytes()).map_err(|err| format!("{:?}", err))?)
         };
-        let flags = u8::try_from(features.flags).map_err(|_| "Invalid output flags: overflowed u8")?;
+        let flags = u16::try_from(features.flags).map_err(|_| "Invalid output flags: overflowed u8")?;
 
         Ok(OutputFeatures::new(
             OutputFeaturesVersion::try_from(
